sadf

+++ incs/colours.h
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 18:25:40 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/11/21 20:45:05 by hmontoya         ###   ########.fr       */
+/*   Updated: 2024/11/22 13:55:10 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,18 +17,18 @@
 
 typedef struct s_color
 {
-	unsigned int r;
-	unsigned int g;
-	unsigned int b;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
 	uint32_t (*normalize)(unsigned int, unsigned int, unsigned int);
 }	t_color;
 
 typedef struct s_acolor
 {
-	unsigned int r;
-	unsigned int g;
-	unsigned int b;
-	unsigned int a;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
+	uint8_t a;
 	uint32_t (*normalize)(unsigned int, unsigned int, unsigned int, unsigned int);
 }	t_acolor;
